+++ packages/node/src/app.module.ts
@@ -1,6 +1,7 @@
 // Copyright 2020-2025 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
+import path from 'node:path';
 import { Module } from '@nestjs/common';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { ScheduleModule } from '@nestjs/schedule';
@@ -8,8 +9,12 @@ import { DbModule, CoreModule, MetaModule } from '@subql/node-core';
 import { ConfigureModule } from './configure/configure.module';
 import { FetchModule } from './indexer/fetch.module';
 
+const kitPath = require.resolve('@solana/kit');
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const { version: ethersSdkVersion } = require('ethers/package.json');
+const { version: solanaSdkVersion } = require(path.resolve(
+  kitPath,
+  '../../package.json',
+));
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { version: packageVersion } = require('../package.json');
 
@@ -23,7 +28,7 @@ const { version: packageVersion } = require('../package.json');
     FetchModule,
     MetaModule.forRoot({
       version: packageVersion,
-      sdkVersion: { name: 'ethers.js', version: ethersSdkVersion },
+      sdkVersion: { name: '@solana/kit', version: solanaSdkVersion },
     }),
   ],
   controllers: [],